Rename infra module to infrastructure
