+++ tests/infrastructure/services/test_pyjwt_token_service.py
@@ -1,8 +1,8 @@
 import jwt
 from pytest import fixture
-from authark.app.models.token import Token
-from authark.app.services.token_service import TokenService
-from authark.infra.crypto.pyjwt_token_service import PyJWTTokenService
+from authark.application.models.token import Token
+from authark.application.services.token_service import TokenService
+from authark.infrastructure.crypto.pyjwt_token_service import PyJWTTokenService
 
 
 def test_pyjwt_token_service_implementation() -> None:
