+++ tests/infrastructure/web/test_main.py
@@ -1,6 +1,6 @@
 from pytest import fixture
 from flask import Flask
-from authark.infra.web.main import main
+from authark.infrastructure.web.main import main
 
 
 def test_web_main() -> None:
